Update the Apple and Google badges
Updated the Apple and Google store badges to the latest ones. New images look more crisp and cleaner.
Also, the new images are directly form the company source. Any changes to the badges in the future should automatically update the repo readme as well.

In general: looks cleaner and helps avoid legal issues related to branding.